+++ src/main/java/com/github/hubbards/algorithms/graph/WeightedGraph.java
@@ -10,18 +10,18 @@ import static com.google.common.base.Preconditions.checkNotNull;
  */
 public abstract class WeightedGraph implements Graph {
     // Default cost of an edge.
-    protected static final float DEFAULT_COST = 0;
+    protected static final double DEFAULT_COST = 0;
 
     /**
      * Adds an edge to this graph with given end-points and cost.
      *
-     * @param tail the name of the head of the given edge, if this graph is
+     * @param tail the name of the tail of the given edge, if this graph is
      *             directed, otherwise the name of one end-point.
-     * @param head the name of the tail of the given edge, if this graph is
+     * @param head the name of the head of the given edge, if this graph is
      *             directed, otherwise the name of the other end-point.
-     * @param cost  the cost of the given edge.
+     * @param cost the cost of the given edge.
      */
-    public abstract void addWeightedEdge(String tail, String head, float cost);
+    public abstract void addWeightedEdge(String tail, String head, double cost);
 
     @Override
     public void addEdge(String tail, String head) {
